requirements.txt
discord.py
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pdfplumber
Pillow
flask
pymupdf
fitz
pdf2image
schedular